requirements.txt
beautifulsoup4==4.13.0
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
idna==3.6
numpy==1.26.4
pandas==2.2.2
python-dateutil==2.9.0.post0
requests==2.31.0
six==1.16.0
soupsieve==2.5
tabula-py==2.9.0
typing_extensions==4.10.0
urllib3==2.2.1
